get scope changes
