+++ frontend/app/layout/SideMenu.tsx
@@ -22,7 +22,8 @@ import { setActiveTab } from 'Duck/search';
 import { Icon } from 'UI';
 import SVG from 'UI/SVG';
 
-import { getScope } from '../duck/user';
+import { getScope } from 'App/duck/user';
+import InitORCard from './InitORCard';
 import SpotToOpenReplayPrompt from './SpotToOpenReplayPrompt';
 import {
   MENU,
@@ -361,35 +362,6 @@ function SideMenu(props: Props) {
       ) : null}
       <SupportModal onClose={() => setSupportOpen(false)} open={supportOpen} />
     </>
-  );
-}
-
-function InitORCard({ onOpenModal }: { onOpenModal: () => void }) {
-  return (
-    <div
-      className={
-        'shadow-sm flex flex-col gap-4 bg-white items-center p-4 mx-auto rounded'
-      }
-      style={{ width: 236 }}
-    >
-      <img src={'/assets/img/init-or.png'} width={200} height={120} />
-      <div className={'font-semibold'}>
-        Discover the full potential of OpenReplay!
-      </div>
-      <div>
-        Empower your product team with essential tools like Session Replay,
-        Product Analytics, Co-Browsing, and more.
-      </div>
-      <Button
-        type="primary"
-        ghost
-        icon={<ArrowRightOutlined />}
-        iconPosition={'end'}
-        onClick={onOpenModal}
-      >
-        Setup OpenReplay Tracker
-      </Button>
-    </div>
   );
 }
